The stopListening wasn't working as expected.
Because Marionette.addEventBinder overwrites the toListen method to automatically _.bind the called method to a context then we can't pass the function in as the third argument to stopListening. This change means that things work as expected BUT it's not really safe. If the same object registered two different listeners to the same event then they would both be unsubscribed by this call rather than just the one we wanted to unsubscribe.

I'm not sure that the proper solution is because Marionette doesn't allow expose the wrapped function. Any ideas?