"Enable example project to be served directly from the examples directory itself.
Created sym-links from backbone.geppetto and its dependencies in to new directory examples/js/lib, and updated the example project's requirejs config to point to them via that location rather than directly to the originals via the repository's root, which is outside the example's web-root when served directly from the examples directory. Adding a pair of sym-links seemed less complex than requiring that it be served from the parent directory, since running a webserver there gives you an index page which leads to examples and text on the modeln github subdomain rather than pointing you straight to the local option.
approach